refactor(Application): 클래스 네이밍 변경 및 애플리케이션 클래스 추가
