+++ src/main/java/lotto/Application.java
@@ -1,12 +1,12 @@
 package lotto;
 
-import lotto.application.LottoApplication;
+import lotto.application.PurchaseApplication;
 import lotto.application.config.AppConfig;
 
 public class Application {
     public static void main(String[] args) {
         AppConfig appConfig = new AppConfig();
-        LottoApplication lottoApplication = new LottoApplication(appConfig);
-        lottoApplication.run();
+        PurchaseApplication purchaseApplication = new PurchaseApplication(appConfig);
+        purchaseApplication.run();
     }
 }
