+++ src/main/java/lotto/application/LottoApplication.java
@@ -1,31 +1,10 @@
 package lotto.application;
 
-import lotto.application.config.AppConfig;
-import lotto.domain.LottoAmount;
-import lotto.domain.PurchasePrice;
-
 public class LottoApplication {
-    private final Printer printer;
-    private final Reader reader;
 
-    public LottoApplication(AppConfig appConfig) {
-        printer = appConfig.printer();
-        reader = appConfig.reader();
-    }
+    private final NumberGenerator numberGenerator;
 
-    public void run() {
-        PurchasePrice purchasePrice = readPurchasePrice();
-        LottoAmount lottoAmount = LottoAmount.from(purchasePrice.getValue());
-        printer.printPurchaseLottoAmount(lottoAmount.getValue());
-    }
-
-    private PurchasePrice readPurchasePrice() {
-        printer.printPurchaseAmountRequest();
-        try {
-            return PurchasePrice.from(reader.read());
-        } catch (IllegalArgumentException e) {
-            printer.print(e.getMessage());
-            return readPurchasePrice();
-        }
+    public LottoApplication(NumberGenerator numberGenerator) {
+        this.numberGenerator = numberGenerator;
     }
 }
